requirements.txt
requests>=2.31.0
pandas>=2.0.0
pyyaml>=6.0.1
duckdb>=0.9.0
tqdm>=4.65.0
psutil>=5.9.0
python-dotenv>=1.0.0